Create MultipleTransformations unit test class; Readme unit test for #transform
